requirements.txt
fastapi
uvicorn
pydantic
pyspark
python-dotenv
pandas